+++ packages/dynamo-config-repository/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./base-repository";
+export * from "./generic-root-config";
+export type { BaseConfig } from "./types";
